[Improve][Transform-V2] Address review: replay-safe event order and honest fixtures for SQL schema change
- Emit translated events by walking the final layout on a simulated sink
  replay, so an AFTER anchor added by the same composite already exists
  and a name needed by an add or rename is freed by an in-place rename
  just before (ADD c FIRST; CHANGE a TO x AFTER c and CHANGE a TO b;
  ADD a FIRST both replay now); rename cycles still fail fast
- Use INT to BIGINT for the derived-column unit test and FLOAT to
  DECIMAL(12,3) for the E2E operand change, because weight * 2 is already
  derived as DOUBLE for a FLOAT operand; start the E2E sink with the
  DOUBLE the transform derives
- Give single incoming events in tests the dialect CDC sources set

+++ seatunnel-e2e/seatunnel-connector-v2-e2e/connector-cdc-mysql-e2e/src/test/resources/ddl/modify_weight_type.sql
@@ -18,13 +18,15 @@
 -- ----------------------------------------------------------------------------------------------------------------
 -- DATABASE:  shop
 -- Changes the type of a column that a SQL transform both projects directly and uses in an expression.
+-- FLOAT to DECIMAL(12,3) also changes the derived type of `weight * 2` from DOUBLE to DECIMAL(12,3); a change
+-- to DOUBLE would not, because the expression is already derived as DOUBLE for a FLOAT operand.
 -- Runs after modify_columns, when products has the columns id, name, description, weight, add_column.
 -- ----------------------------------------------------------------------------------------------------------------
 CREATE DATABASE IF NOT EXISTS `shop`;
 
 use shop;
 
-alter table products modify weight double;
+alter table products modify weight decimal(12,3);
 
 insert into products
 values (182,"scooter","Small 2-wheel scooter",3.14,1),
